Just export properties we want to test
